requirements.txt
neptune-cli
steppy-toolkit
steppy==0.1.4
attrdict==2.0.0
category_encoders==1.2.6
click==6.7
lightgbm==2.1.1
numpy==1.14.3
pandas==0.23.1
scikit_learn==0.19.1
PyYAML==3.12
xgboost==0.72.1
